+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Change Log
 
 ## [Unreleased]
+### Changed
+- Use useIsomorphicLayoutEffect hack to avoid warning in SSR
 
 ## [3.8.0] - 2019-11-26
 ### Changed